+++ rest-openehr/src/main/java/org/ehrbase/rest/openehr/controller/admin/AdminApiConfiguration.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2019 Axel Siebert (Vitasystems GmbH) and Hannover Medical School.
+ *
+ * This file is part of project EHRbase
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.ehrbase.rest.openehr.controller.admin;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Component
+@ConfigurationProperties(prefix = "admin-api")
+public class AdminApiConfiguration {
+
+    private boolean active;
+
+    private boolean allowDeleteAll;
+
+    public boolean getActive() {
+        return this.active;
+    }
+
+    public void setActive(boolean active) {
+        this.active = active;
+    }
+
+    public boolean getAllowDeleteAll() {
+        return this.allowDeleteAll;
+    }
+
+    public void setAllowDeleteAll(boolean allowDeleteAll) {
+        this.allowDeleteAll = allowDeleteAll;
+    }
+}