+++ rest-openehr/src/main/java/org/ehrbase/rest/openehr/controller/admin/OpenehrAdminController.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright (c) 2020 Vitasystems GmbH and Jake Smolka (Hannover Medical School).
+ *
+ * This file is part of project EHRbase
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.ehrbase.rest.openehr.controller.admin;
+
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
+import io.swagger.annotations.ResponseHeader;
+import org.ehrbase.response.openehr.admin.AdminStatusResponseData;
+import org.ehrbase.rest.openehr.controller.BaseController;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@Api(tags = {"Admin", "Heartbeat"})
+@ConditionalOnProperty(prefix = "admin-api", name = "active")
+@RestController
+@RequestMapping(path = "/rest/openehr/v1/admin", produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
+public class OpenehrAdminController extends BaseController {
+
+    @GetMapping(path = "/status")
+    @ApiResponses(value = {
+            @ApiResponse(
+                    code = 200,
+                    message = "Admin API resources available and user has permission to access.",
+                    responseHeaders = {
+                            @ResponseHeader(
+                                    name = CONTENT_TYPE,
+                                    description = RESP_CONTENT_TYPE_DESC,
+                                    response = MediaType.class
+                            )
+                    }
+            ),
+            @ApiResponse(code = 401, message = "Client credentials are invalid or have expired."),
+            @ApiResponse(code = 403, message = "Client has no access permission since the admin role is missing.")
+    })
+    public ResponseEntity<AdminStatusResponseData> getStatus() {
+
+        return ResponseEntity.ok().body(
+                new AdminStatusResponseData("EHRbase Admin API available and you have permission to access it")
+        );
+    }
+}